% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger_graph.R
\name{ledger_graph}
\alias{ledger_graph}
\title{Graph your net worth}
\usage{
ledger_graph(data, start = NA, end = Sys.Date(), separate = FALSE,
  aggregate = "day", cumulative = FALSE)
}
\arguments{
\item{data}{the data to graph}

\item{start}{the start date (e.g. "2017-01-01"), defaults to the start of the
series}

\item{end}{the end date (e.g. "2017-12-31"), defaults to today}
}
\value{
a "ggplot2" graph object
}
\description{
Helper function to graph net worth over time
}
\examples{
demo <- system.file("extdata", "demo.ledger", package="ledgerReports")
ledge <- read_ledger("'^assets' '^exp' -X $", file = demo)
ledger_graph(ledge, end = "2012-01-01")
}
