% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_worth.R
\name{net_worth}
\alias{net_worth}
\title{Graph your net worth}
\usage{
net_worth(data, start = NA, end = Sys.Date())
}
\arguments{
\item{data}{the data to graph}

\item{start}{the start date (e.g. "2017-01-01"), defaults to the start of the
series}

\item{end}{the end date (e.g. "2017-12-31"), defaults to today}
}
\value{
a "ggplot2" graph object
}
\description{
Helper function to graph net worth over time
}
