% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ledger.R
\name{read_ledger}
\alias{read_ledger}
\title{Read in from a ledger file}
\usage{
read_ledger(report_query, file = Sys.getenv("LEDGER_FILE"))
}
\arguments{
\item{report_query}{report query to pass to ledger csv}

\item{file}{the ledger file to read from}
}
\value{
a data.frame object containing 6 columns - date, payee, account,
  commodity, amount, and note
}
\description{
\code{read_ledger} reads data from a ledger file using \code{ledger csv}
}
\examples{
demo <- system.file("extdata", "demo.ledger", package="ledgerReports")
read_ledger("'^assets' '^liab' -X $", file = demo)
}
