% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{time_aggregate}
\alias{time_aggregate}
\title{Aggregate ledger data.frames by time blocks}
\usage{
time_aggregate(data, cumulative = FALSE, aggregate = "day")
}
\arguments{
\item{data, }{the data file (usually produced by \code{read_ledger})}

\item{cumulative}{should a cumulative sum be added? defaults to FALSE}

\item{aggregate}{string indicating what level to aggregate to. See \code{lubridate::floor_date} for supported units. Defaults to "day".}
}
\value{
a tibble (data.frame)
}
\description{
Aggregate data by time blocks. Currently only supports daily.
}
\examples{
demo <- system.file("extdata", "demo.ledger", package="ledgerReports")
ledge <- read_ledger("'^assets' '^liab' -X $", file = demo)
time_aggregate(ledge, cumulative = TRUE)
}
